Common file loan_common.js for loan and loan_application
