+++ erpnext/hr/doctype/loan/loan.json
@@ -57,7 +57,7 @@
    "ignore_user_permissions": 0, 
    "ignore_xss_filter": 0, 
    "in_filter": 0, 
-   "in_global_search": 1, 
+   "in_global_search": 0, 
    "in_list_view": 0, 
    "in_standard_filter": 1, 
    "label": "Applicant", 
@@ -84,7 +84,7 @@
    "collapsible": 0, 
    "columns": 0, 
    "fieldname": "applicant_name", 
-   "fieldtype": "Read Only", 
+   "fieldtype": "Data", 
    "hidden": 0, 
    "ignore_user_permissions": 0, 
    "ignore_xss_filter": 0, 
@@ -99,7 +99,7 @@
    "precision": "", 
    "print_hide": 0, 
    "print_hide_if_no_value": 0, 
-   "read_only": 0, 
+   "read_only": 1, 
    "remember_last_selected_value": 0, 
    "report_hide": 0, 
    "reqd": 0, 
@@ -1009,7 +1009,7 @@
  "issingle": 0, 
  "istable": 0, 
  "max_attachments": 0, 
- "modified": "2018-02-26 08:52:22.056246", 
+ "modified": "2018-02-26 13:39:49.906153", 
  "modified_by": "Administrator", 
  "module": "HR", 
  "name": "Loan",
