+++ erpnext/hr/doctype/loan/loan_dashboard.py
@@ -1,4 +1,5 @@
 from frappe import _
+import frappe
 
 def get_data():
 	return {
